Refactor Disk/Network views and fix rendering artifacts.
- Refactored DiskView and NetworkView to use timeout-based semaphore acquisition instead of Pulse-driven polling. This improves responsiveness and decouples data collection from the UI thread.
- Fixed drawing artifacts in ActivityGraphView by clipping partial updates.
- Standardized CircularBuffer size type to uint32.

+++ SystemMonitor/CircularBuffer.h
@@ -15,7 +15,7 @@
 template<typename Type>
 class CircularBuffer {
 public:
-	CircularBuffer(size_t size)
+	CircularBuffer(uint32 size)
 		:
 		fFirst(0),
 		fIn(0),
@@ -80,7 +80,7 @@ public:
 		return fBuffer != NULL ? B_OK : B_NO_MEMORY;
 	}
 
-	status_t SetSize(size_t size)
+	status_t SetSize(uint32 size)
 	{
 		if (fSize == size)
 			return B_OK;
